requirements.txt
django>=6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
python-slugify>=8.0
Faker>=28.0
